ADVC-31 Cria componente datalayer
